@@ -1,9 +1,22 @@
-import React, { ReactNode } from 'react';
-import { Components } from '@episerver/spa-core';
+import React, { useState, useEffect } from 'react';
+import { ComponentTypes, useEpiserver } from '@episerver/spa-core';
 
-export default class MoseyLoader extends Components.Spinner {
-    render() : ReactNode
-    {
+export const MoseyLoader : ComponentTypes.Spinner = (props) => {
+    var ctx = useEpiserver();
+	var timeout = props.timeout || ctx.config().spinnerTimeout || 0;
+	var [isVisible, setIsVisible] = useState<boolean>(timeout === 0);
+	if (ctx.config().enableSpinner !== true) return null;
+
+	useEffect(() => {
+		if (timeout === 0) return;
+		setIsVisible(false);
+		const timeoutHandle = setTimeout(() => { setIsVisible(true); }, timeout);
+		return () => {
+			clearTimeout(timeoutHandle)
+		}
+	}, []);
+
+	if (isVisible) {
         return <div className="mosey-loader alert alert-primary text-center m-3" role="alert">
             <img src="/globalassets/_epihealth/logo_frontlineservices.png" alt="" className="img-fluid m-3" style={ { maxWidth: "25rem"} } />
             <hr/>
@@ -13,4 +26,7 @@ export default class MoseyLoader extends Components.Spinner {
             </p>
         </div>
     }
+    return null;
 }
+
+export default MoseyLoader;
