@@ -1,28 +1,58 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
+using System;
 using System.Web;
 
 namespace Foundation.SpaViewEngine.JsInterop.Models
 {
+    /// <summary>
+    /// Simple wrapper class to create a JavaScript Location object from a 
+    /// .Net System.Uri;
+    /// </summary>
     public class Location : JavaScriptObject
     {
-        protected Uri CurrentUri => HttpContext.Current.Request.Url;
+        protected readonly Uri CurrentUri;
 
-        //Disable the method naming convention message, as we're implementing a JavaScript API here.
-#pragma warning disable IDE1006
-        public string href => CurrentUri.AbsoluteUri;
+        public Location() : this(HttpContext.Current.Request.Url) { }
 
-        public string hostname => CurrentUri.Host;
+        public Location(Uri uri)
+        {
+            CurrentUri = uri;
+        }
 
-        public string pathname => CurrentUri.AbsolutePath;
+        [JsonProperty(PropertyName = "href")]
+        public string Href => CurrentUri.AbsoluteUri;
 
-        public string protocol => CurrentUri.Scheme;
-#pragma warning restore IDE1006
+        [JsonProperty(PropertyName = "hostname")]
+        public string Hostname => CurrentUri.Host;
 
-        /*
-        window.location.href returns the href(URL) of the current page
-        window.location.hostname returns the domain name of the web host
-        window.location.pathname returns the path and filename of the current page
-        window.location.protocol returns the web protocol used (http: or https:)
-        */
+        [JsonProperty(PropertyName = "host")]
+        public string Host => CurrentUri.Authority;
+
+        [JsonProperty(PropertyName = "origin")]
+        public string Origin => CurrentUri.Scheme + "//" + CurrentUri.Authority;
+
+        [JsonProperty(PropertyName = "pathname")]
+        public string Pathname => CurrentUri.AbsolutePath;
+
+        [JsonProperty(PropertyName = "protocol")]
+        public string Protocol => CurrentUri.Scheme;
+
+        [JsonProperty(PropertyName = "search")]
+        public string Search => CurrentUri.Query;
+
+        public override string ToString() => Href;
+
+#pragma warning disable IDE1006 // Naming conventions between JavaScript and .Net are different
+        public virtual void replace()
+        {
+            throw new NotImplementedException();
+        }
+        public virtual void reload()
+        {
+            throw new NotImplementedException();
+        }
+
+#pragma warning restore IDE1006 // Naming conventions between JavaScript and .Net are different
     }
 }
