@@ -1,4 +1,7 @@
-﻿namespace Foundation.SpaViewEngine.JsInterop.Models
+﻿
+using System.Collections.Generic;
+
+namespace Foundation.SpaViewEngine.JsInterop.Models
 {
     public class SSRResponse
     {
@@ -10,5 +13,7 @@
         public string Script { get; set; } = "";
         public string Style { get; set; } = "";
         public string BodyAttributes { get; set; } = "";
+        public int StatusCode { get; set; } = 200;
+        public Dictionary<string, string> Headers { get; set; } = new Dictionary<string, string>();
     }
 }
