@@ -1,6 +1,6 @@
 ﻿@using Foundation.Features.Blocks.FacebookBlock
 
-﻿@model IBlockViewModel<FacebookBlock>
+@model IBlockViewModel<FacebookBlock>
 
 <div id="fb-root"></div>
 <script async defer crossorigin="anonymous" src="https://connect.facebook.net/en_US/sdk.js#xfbml=1&version=v3.3"></script>
