@@ -4,6 +4,7 @@ using EPiServer.Filters;
 using EPiServer.Framework.DataAnnotations;
 using EPiServer.Web.Mvc;
 using Foundation.Cms;
+using Foundation.Features.Folder;
 using Geta.EpiCategories;
 using System.Collections.Generic;
 using System.Linq;
@@ -25,7 +26,7 @@ namespace Foundation.Features.Blocks.PageListBlock
         public override ActionResult Index(PageListBlock currentBlock)
         {
             var pages = FindPages(currentBlock);
-
+            pages = pages.Where(x => x.PageTypeName != typeof(FolderPage).Name);
             pages = Sort(pages, currentBlock.SortOrder);
 
             if (currentBlock.Count > 0)
@@ -41,7 +42,7 @@ namespace Foundation.Features.Blocks.PageListBlock
             ViewData.GetEditHints<PageListBlockViewModel, PageListBlock>()
                 .AddConnection(x => x.Heading, x => x.Heading);
 
-            return PartialView("~/Features/Blocks/PageListBlock/View/PageListBlock.cshtml", model);
+            return PartialView("~/Features/Blocks/PageListBlock/Views/PageListBlock.cshtml", model);
         }
 
         private IEnumerable<PageData> FindPages(PageListBlock currentBlock)
