@@ -20,66 +20,75 @@ namespace Foundation.Features.Blocks.TeaserBlock
     [ImageUrl("~/assets/icons/cms/blocks/CMS-icon-block-26.png")]
     public class TeaserBlock : FoundationBlockData, IDashboardItem
     {
+        #region Content
+        [CultureSpecific]
+        [Display(GroupName = SystemTabNames.Content, Order = 20)]
+        public virtual string Description { get; set; }
+
+        [Display(GroupName = SystemTabNames.Content, Order = 21)]
+        public virtual PageReference Link { get; set; }
+        #endregion
+
+        #region Header
         [CultureSpecific]
         [Required(AllowEmptyStrings = false)]
-        [Display(GroupName = SystemTabNames.Content, Order = 10)]
+        [Display(Name = "Heading text", GroupName = TabNames.Header, Order = 10)]
         public virtual string Heading { get; set; }
 
-        [Display(Name = "Heading size", GroupName = SystemTabNames.Content, Order = 11)]
+        [Display(Name = "Heading size", GroupName = TabNames.Header, Order = 11)]
         public virtual int HeadingSize { get; set; }
 
         [SelectOne(SelectionFactoryType = typeof(TeaserBlockHeadingStyleSelectionFactory))]
-        [Display(Name = "Heading style", GroupName = SystemTabNames.Content, Order = 12)]
+        [Display(Name = "Heading style", GroupName = TabNames.Header, Order = 12)]
         public virtual string HeadingStyle { get; set; }
 
         [ClientEditor(ClientEditingClass = "foundation/editors/ColorPicker")]
-        [Display(Name = "Heading color", GroupName = SystemTabNames.Content, Order = 13)]
+        [Display(Name = "Heading color", GroupName = TabNames.Header, Order = 13)]
         public virtual string HeadingColor
         {
             get { return this.GetPropertyValue(page => page.HeadingColor) ?? "#000000ff"; }
             set { this.SetPropertyValue(page => page.HeadingColor, value); }
         }
+        #endregion
 
+        #region Text
         [CultureSpecific]
-        [Display(GroupName = SystemTabNames.Content, Order = 20)]
-        public virtual string Description { get; set; }
-
-        [CultureSpecific]
-        [Display(GroupName = SystemTabNames.Content, Order = 30)]
+        [Display(GroupName = TabNames.Text, Order = 30)]
         public virtual XhtmlString Text { get; set; }
-
-        [CultureSpecific]
-        [UIHint(UIHint.Image)]
-        [Display(GroupName = SystemTabNames.Content, Order = 40)]
-        public virtual ContentReference Image { get; set; }
-
-        [Range(1, 100, ErrorMessage = "Set image width from 1 to 100")]
-        [Display(Name = "Image size (%)", GroupName = SystemTabNames.Content, Order = 41)]
-        public virtual int ImageSize { get; set; }
-
-        [CultureSpecific]
-        [UIHint(UIHint.Image)]
-        [Display(Name = "Second Image", GroupName = SystemTabNames.Content, Order = 45)]
-        public virtual ContentReference SecondImage { get; set; }
-
-        [Range(1, 100, ErrorMessage = "Set image width from 1 to 100")]
-        [Display(Name = "Image size (%)", GroupName = SystemTabNames.Content, Order = 46)]
-        public virtual int SecondImageSize { get; set; }
-
-        [Display(GroupName = SystemTabNames.Content, Order = 90)]
-        public virtual PageReference Link { get; set; }
-
         [ClientEditor(ClientEditingClass = "foundation/editors/ColorPicker")]
-        [Display(Name = "Text color", GroupName = SystemTabNames.Content, Order = 50)]
+        [Display(Name = "Text color", GroupName = TabNames.Text, Order = 50)]
         public virtual string TextColor
         {
             get { return this.GetPropertyValue(page => page.TextColor) ?? "#000000ff"; }
             set { this.SetPropertyValue(page => page.TextColor, value); }
         }
+        #endregion
 
+        #region Image
+        [CultureSpecific]
+        [UIHint(UIHint.Image)]
+        [Display(GroupName = TabNames.Image, Order = 40)]
+        public virtual ContentReference Image { get; set; }
+
+        [Range(1, 100, ErrorMessage = "Set image width from 1 to 100")]
+        [Display(Name = "Image size (%)", GroupName = TabNames.Image, Order = 41)]
+        public virtual int ImageSize { get; set; }
+
+        [CultureSpecific]
+        [UIHint(UIHint.Image)]
+        [Display(Name = "Second Image", GroupName = TabNames.Image, Order = 45)]
+        public virtual ContentReference SecondImage { get; set; }
+
+        [Range(1, 100, ErrorMessage = "Set image width from 1 to 100")]
+        [Display(Name = "Image size (%)", GroupName = TabNames.Image, Order = 46)]
+        public virtual int SecondImageSize { get; set; }
+        #endregion
+
+        #region Style
         [SelectOne(SelectionFactoryType = typeof(TeaserBlockHeightStyleSelectionFactory))]
         [Display(Name = "Height", GroupName = TabNames.BlockStyling, Order = 100)]
         public virtual string Height { get; set; }
+        #endregion
 
         public override void SetDefaultValues(ContentType contentType)
         {
