@@ -28,8 +28,8 @@
             </div>
         </div>
         <div id="blog-list">
-            @Html.Partial("_BlogList", Model)
-            @Html.Partial("_Paging", Model)
+            @Html.Partial("Views/_BlogList", Model)
+            @Html.Partial("Views/_Paging", Model)
         </div>
     </div>
 </div>
