@@ -9,7 +9,7 @@
     </h4>
     <ul class="portfolio-cats">
         <li>
-            <span style="width:100%; text-align:center">
+            <span class="w-100 text-center">
                 <input id="slider-range" data-slider-id='slider-rangeData' type="text" style="width:95%;" />
             </span>
         </li>
