@@ -1,4 +1,6 @@
 ﻿@using ImageProcessor.Web.Episerver
+@using Foundation.Features.Locations.TagPage
+
 @model TagsCarouselViewModel
 
 @if (Model.Items != null)
@@ -20,7 +22,7 @@
                     <a href="@Url.ContentUrl(item.value.ItemURL)">
                         <picture class="img-fluid">
                             <source data-srcset="@Html.ProcessImage(@Url.ContentUrl(item.value.Image)).Resize(1440, 820)?format=webp" type="image/webp">
-                            <img class="img-fluid lazyload" data-src="@Html.ProcessImage(@Url.ContentUrl(item.value.Image)).Resize(1440, 820)"/>
+                            <img class="img-fluid lazyload" data-src="@Html.ProcessImage(@Url.ContentUrl(item.value.Image)).Resize(1440, 820)" />
                         </picture>
                     </a>
                     @if (!string.IsNullOrWhiteSpace(item.value.Heading))
