@@ -1,4 +1,6 @@
-﻿@model TagsViewModel
+﻿@using Foundation.Features.Locations.TagPage
+
+@model TagsViewModel
 
 <section class="container">
     <div class="row">
