@@ -1,4 +1,4 @@
-﻿﻿@using EPiServer.SpecializedProperties;
+﻿@using EPiServer.SpecializedProperties;
 
 @using EPiServer.SpecializedProperties
 @model Foundation.Features.Settings.LayoutSettings
