@@ -39,5 +39,5 @@ else
 
 @RenderSection("AdditionalScripts", required: false)
 @Html.RequiredClientResources("Footer")
-@Html.RenderExtendedScripts(Model.CurrentContent)
+@Html.RenderFooterScripts(Model.CurrentContent)
 @Model.SchemaMarkup
