@@ -1,4 +1,4 @@
-﻿﻿@model IContentViewModel<IContent>
+﻿@model IContentViewModel<IContent>
 
 <!DOCTYPE html>
 <html lang="en" class="@(PageEditing.PageIsInEditMode ? "editMode" : "")">
@@ -24,6 +24,7 @@
     @RenderSection("AdditionalStyles", required: false)
     @Html.RequiredClientResources("Header")
     @Html.RenderExtendedCss(Model.CurrentContent)
+    @Html.RenderHeaderScripts(Model.CurrentContent)
 </head>
 <body>
     @RenderBody()
