@@ -6,7 +6,9 @@ using EPiServer.Web;
 using EPiServer.Web.Mvc.Html;
 using EPiServer.Web.Routing;
 using Foundation.Cms.Extensions;
+using Foundation.Cms.Settings;
 using Foundation.Features.Home;
+using Foundation.Features.Settings;
 using Foundation.Features.Shared;
 using System;
 using System.Collections.Generic;
@@ -23,7 +25,7 @@ namespace Foundation.Helpers
     {
         private const string _cssFormat = "<link href=\"{0}\" rel=\"stylesheet\" />";
         private const string _scriptFormat = "<script src=\"{0}\"></script>";
-        private const string _metaFormat = "<meta property=\"{0}\" content=\"{1}\" />";
+        private const string _metaFormat = "<meta name=\"{0}\" property=\"{0}\" content=\"{1}\" />";
 
         private static readonly Lazy<IContentLoader> _contentLoader =
             new Lazy<IContentLoader>(() => ServiceLocator.Current.GetInstance<IContentLoader>());
@@ -33,6 +35,9 @@ namespace Foundation.Helpers
 
         private static readonly Lazy<IPermanentLinkMapper> _permanentLinkMapper =
            new Lazy<IPermanentLinkMapper>(() => ServiceLocator.Current.GetInstance<IPermanentLinkMapper>());
+
+        private static readonly Lazy<ISettingsService> _settingsService =
+           new Lazy<ISettingsService>(() => ServiceLocator.Current.GetInstance<ISettingsService>());
 
         public static MvcHtmlString RenderExtendedCss(this HtmlHelper helper, IContent content)
         {
@@ -63,7 +68,7 @@ namespace Foundation.Helpers
             return new MvcHtmlString(outputCss.ToString());
         }
 
-        public static MvcHtmlString RenderExtendedScripts(this HtmlHelper helper, IContent content)
+        public static MvcHtmlString RenderHeaderScripts(this HtmlHelper helper, IContent content)
         {
             if (content == null || ContentReference.StartPage == PageReference.EmptyReference || !(content is FoundationPageData sitePageData))
             {
@@ -71,22 +76,75 @@ namespace Foundation.Helpers
             }
 
             var outputScript = new StringBuilder(string.Empty);
-            var startPage = _contentLoader.Value.Get<HomePage>(ContentReference.StartPage);
 
-            // Extended Javascript file
-            AppendFiles(startPage.ScriptFiles, outputScript, _scriptFormat);
-            if (!(sitePageData is HomePage))
+            // Injection Hierarchically Javascript
+            var settings = _settingsService.Value.GetSiteSettings<ScriptInjectionSettings>();
+            if (settings != null && settings.HeaderScripts != null)
             {
-                AppendFiles(sitePageData.ScriptFiles, outputScript, _scriptFormat);
+                foreach (var script in settings.HeaderScripts)
+                {
+                    var pages = _contentLoader.Value.GetDescendents(script.ScriptRoot);
+                    if (pages.Any(x => content.ContentLink.ID == x.ID) || content.ContentLink.ID == script.ScriptRoot.ID)
+                    {
+                        // Script Files
+                        AppendFiles(script.ScriptFiles, outputScript, _scriptFormat);
+
+                        // External Javascript
+                        if (!string.IsNullOrWhiteSpace(script.ExternalScripts))
+                        {
+                            outputScript.AppendLine(script.ExternalScripts);
+                        }
+
+                        // Inline Javascript
+                        if (!string.IsNullOrWhiteSpace(script.InlineScripts))
+                        {
+                            outputScript.AppendLine("<script type=\"text/javascript\">");
+                            outputScript.AppendLine(!string.IsNullOrWhiteSpace(script.InlineScripts) ? script.InlineScripts : "");
+                            outputScript.AppendLine("</script>");
+                        }
+                    }
+                }
             }
 
-            // Inline Javascript
-            if (!string.IsNullOrWhiteSpace(startPage.Scripts) || !string.IsNullOrWhiteSpace(sitePageData.Scripts))
+            return new MvcHtmlString(outputScript.ToString());
+        }
+
+        public static MvcHtmlString RenderFooterScripts(this HtmlHelper helper, IContent content)
+        {
+            if (content == null || ContentReference.StartPage == PageReference.EmptyReference || !(content is FoundationPageData sitePageData))
             {
-                outputScript.AppendLine("<script type=\"text/javascript\">");
-                outputScript.AppendLine(!string.IsNullOrWhiteSpace(startPage.Scripts) ? startPage.Scripts : "");
-                outputScript.AppendLine(!string.IsNullOrWhiteSpace(sitePageData.Scripts) && !(sitePageData is HomePage) ? sitePageData.Scripts : "");
-                outputScript.AppendLine("</script>");
+                return new MvcHtmlString("");
+            }
+
+            var outputScript = new StringBuilder(string.Empty);
+
+            // Injection Hierarchically Javascript
+            var settings = _settingsService.Value.GetSiteSettings<ScriptInjectionSettings>();
+            if (settings != null && settings.FooterScripts != null)
+            {
+                foreach (var script in settings.FooterScripts)
+                {
+                    var pages = _contentLoader.Value.GetDescendents(script.ScriptRoot);
+                    if (pages.Any(x => content.ContentLink.ID == x.ID) || content.ContentLink.ID == script.ScriptRoot.ID)
+                    {
+                        // Script Files
+                        AppendFiles(script.ScriptFiles, outputScript, _scriptFormat);
+
+                        // External Javascript
+                        if (!string.IsNullOrWhiteSpace(script.ExternalScripts))
+                        {
+                            outputScript.AppendLine(script.ExternalScripts);
+                        }
+
+                        // Inline Javascript
+                        if (!string.IsNullOrWhiteSpace(script.InlineScripts))
+                        {
+                            outputScript.AppendLine("<script type=\"text/javascript\">");
+                            outputScript.AppendLine(!string.IsNullOrWhiteSpace(script.InlineScripts) ? script.InlineScripts : "");
+                            outputScript.AppendLine("</script>");
+                        }
+                    }
+                }
             }
 
             return new MvcHtmlString(outputScript.ToString());
@@ -133,6 +191,20 @@ namespace Foundation.Helpers
                 outputString.AppendLine(map == null
                     ? string.Format(formatString, item.GetMappedHref())
                     : string.Format(formatString, _urlResolver.Value.GetUrl(map.ContentReference)));
+            }
+        }
+
+        private static void AppendFiles(IList<ContentReference> files, StringBuilder outputString, string formatString)
+        {
+            if (files == null || files.Count <= 0)
+            {
+                return;
+            }
+
+            foreach (var item in files.Where(item => !string.IsNullOrEmpty(_urlResolver.Value.GetUrl(item))))
+            {
+                var url = _urlResolver.Value.GetUrl(item);
+                outputString.AppendLine(string.Format(formatString, url));
             }
         }
 
