@@ -13,7 +13,8 @@ namespace Foundation.Infrastructure.Display
             "~/Features/Blocks/{1}/{0}.cshtml",
             "~/Features/Shared/Views/{0}.cshtml",
             "~/Features/Shared/Views/{1}/{0}.cshtml",
-            "~/Features/Shared/Views/Header/{0}.cshtml"
+            "~/Features/Shared/Views/Header/{0}.cshtml",
+            "~/Cms/Views/{1}/{0}.cshtml",
         };
 
         public FeaturesViewEngine()
@@ -23,7 +24,8 @@ namespace Foundation.Infrastructure.Display
             var featureFolders = new[]
             {
                 "~/Features/%1/{1}/{0}.cshtml",
-                "~/Features/%1/{0}.cshtml"
+                "~/Features/%1/{0}.cshtml",
+                "~/Cms/%1/{1}/{0}.cshtml"
             };
 
             featureFolders = featureFolders.Union(AdditionalPartialViewFormats).ToArray();
