@@ -1,17 +1,48 @@
-import { ContentDelivery, Taxonomy } from '@episerver/spa-core';
+import { ContentDelivery, ServerSideRendering, Taxonomy } from '@episerver/spa-core';
+
+type ISettingsService = {
+    GetSiteSettings<T extends Taxonomy.IContent = Taxonomy.IContent>(container: string) : T | undefined;
+}
+
+type SettingApiContext = {
+    Settings : { [container: string ] : Taxonomy.IContent }
+}
 
 export default class SettingsApi
 {
     public serviceEndpoint : Readonly<string> = 'api/foundation/v1/settings';
     protected _repo : Readonly<ContentDelivery.IIContentRepositoryV2>;
     protected _api : Readonly<ContentDelivery.IContentDeliveryAPI_V2>;
+    protected _ctx : Readonly<ServerSideRendering.IAccessor>;
 
     public constructor (
         contentDeliveryApi: ContentDelivery.IContentDeliveryAPI_V2,
-        contentRepositoryApi: ContentDelivery.IIContentRepositoryV2
+        contentRepositoryApi: ContentDelivery.IIContentRepositoryV2,
+        serverContext: ServerSideRendering.IAccessor
     ) {
         this._api = contentDeliveryApi;
         this._repo = contentRepositoryApi;
+        this._ctx = serverContext;
+    }
+
+    /**
+     * Get a settings container, either during server side rendering or from the initial context delivered by the
+     * server side rendering.
+     * 
+     * @param       container 
+     * @returns     The settings container, or undefined if not found
+     */
+    public getContainerOnServer<T extends Taxonomy.IContent = Taxonomy.IContent>(container: string) : T | undefined
+    {
+        const ssrPropName = container.toLowerCase();
+        let settingsContainer  = this._ctx.getProp<T>(ssrPropName);
+        if (!settingsContainer && this._ctx.IsServerSideRendering) {
+            const ISettingsService = (this._ctx as ServerSideRendering.DotNetAccessor).getEpiserverService<ISettingsService>("Foundation.Cms.Settings.ISettingsService");
+            settingsContainer = ISettingsService.GetSiteSettings<T>(container);
+            this._ctx.setProp(ssrPropName, settingsContainer);
+            settingsContainer = (this._ctx as ServerSideRendering.DotNetAccessor).makeSafe(settingsContainer);
+        }
+        return settingsContainer;
     }
 
     public async listContainers(site ?: Taxonomy.Website) : Promise<string[]>
@@ -19,7 +50,7 @@ export default class SettingsApi
         site = site || await this._repo.getCurrentWebsite();
         // @ToDo: Implement local storage caching...
         let url = this.serviceEndpoint + '/' + site.id;
-        return this._api.raw<string[]>(url).then(r => r[0]);
+        return this._api.raw<string[]>(url).then(r => ContentDelivery.isNetworkError(r[0]) ? [] : r[0]);
     }
 
     public async getContainer<T extends Taxonomy.IContent = Taxonomy.IContent>(container: string, site ?: Taxonomy.Website) : Promise<T | null>
