@@ -29,37 +29,37 @@ export default class ImageMediaData extends ComponentTypes.AbstractComponent<Ima
 
     public render() : ReactNode {
         let cssClass : string = "img-fluid"
-        if (this.props.className !== undefined) {
-            cssClass = cssClass + " " + this.props.className
-        }
-        if (this.props.data.url !== null) {
-            let imgUrl: URL = new URL(Core.DefaultContext.getEpiserverUrl(this.props.data.url));
-            imgUrl.searchParams.set("upscale", "false"); //Never upscale an image, leave that to the browser
-            imgUrl.searchParams.set("quality", "85"); //Lossy processing
-            let webpUrl : URL = new URL(Core.DefaultContext.getEpiserverUrl(this.props.data.url));
-            webpUrl.searchParams.set("upscale", "false"); //Never upscale an image, leave that to the browser
-            webpUrl.searchParams.set("format", "webp"); //Format as WebP
-            webpUrl.searchParams.set("quality", "85"); //Lossy processing
+        let altText : string = this.props.data?.altText?.value || (typeof (this.props.data?.name) == 'string' ? this.props.data.name : this.props.data?.name?.value) || "image";
+        if (this.props.className !== undefined)
+            cssClass = cssClass + " " + this.props.className;
+        
+        if (!this.props.data.url)
+            return <span className={ cssClass.trim() + " no-img" }/>;
 
-            let sources: ReactNodeArray = [];
-            let keyPrefix : string = `img-${ this.props.data.contentLink.id }-`;
-            for (let size in this.breakPoints)
-            {
-                imgUrl.searchParams.set("width", this.breakPoints[size].imgWidth);
-                webpUrl.searchParams.set("width", this.breakPoints[size].imgWidth);
-                if (size == "img") {
-                    sources.push(<img src={imgUrl.href} className={ cssClass.trim() }  key={ keyPrefix + size } />);
-                    sources.push(<source media={ this.breakPoints[size].cssMedia } srcSet={ webpUrl.href }  type="image/webp" key={ keyPrefix + size + '-webp'}/>);
-                } else {
-                    sources.push(<source media={ this.breakPoints[size].cssMedia } srcSet={ imgUrl.href }  key={ keyPrefix + size }/>);
-                    sources.push(<source media={ this.breakPoints[size].cssMedia } srcSet={ webpUrl.href }  type="image/webp" key={ keyPrefix + size + '-webp'} />);
-                }
+        let imgUrl: URL = new URL(Core.DefaultContext.getEpiserverUrl(this.props.data.url));
+        imgUrl.searchParams.set("upscale", "false"); //Never upscale an image, leave that to the browser
+        imgUrl.searchParams.set("quality", "85"); //Lossy processing
+        let webpUrl : URL = new URL(Core.DefaultContext.getEpiserverUrl(this.props.data.url));
+        webpUrl.searchParams.set("upscale", "false"); //Never upscale an image, leave that to the browser
+        webpUrl.searchParams.set("format", "webp"); //Format as WebP
+        webpUrl.searchParams.set("quality", "85"); //Lossy processing
+
+        let sources: ReactNodeArray = [];
+        let keyPrefix : string = `img-${ this.props.data.contentLink.id }-`;
+        for (let size in this.breakPoints)
+        {
+            imgUrl.searchParams.set("width", this.breakPoints[size].imgWidth);
+            webpUrl.searchParams.set("width", this.breakPoints[size].imgWidth);
+            if (size == "img") {
+                sources.push(<source media={ this.breakPoints[size].cssMedia } srcSet={ webpUrl.href }  type="image/webp" key={ keyPrefix + size + '-webp'}/>);
+                sources.push(<img src={imgUrl.href} className={ cssClass.trim() }  key={ keyPrefix + size } alt={ altText }/>);
+            } else {
+                sources.push(<source media={ this.breakPoints[size].cssMedia } srcSet={ webpUrl.href }  type="image/webp" key={ keyPrefix + size + '-webp'} />);
+                sources.push(<source media={ this.breakPoints[size].cssMedia } srcSet={ imgUrl.href }  key={ keyPrefix + size }/>);
             }
-            //console.log(this.props.data);
-            return <picture>
-                {sources}
-            </picture>
         }
-        return <span className={ cssClass.trim() + " no-img" }/>;
+        return <picture>
+            {sources}
+        </picture>
     }
 }
