@@ -3,28 +3,30 @@ import { Taxonomy, Core, Loaders } from '@episerver/spa-core';
 export default class TypeMapper extends Loaders.BaseTypeMapper {
   protected map : { [type: string]: Loaders.TypeInfo } = {
     'CookieDropBlock': {dataModel: 'CookieDropBlockData',instanceModel: 'CookieDropBlockType'},
+    'ResetPasswordMailPage': {dataModel: 'ResetPasswordMailPageData',instanceModel: 'ResetPasswordMailPageType'},
+    'PersonPage': {dataModel: 'PersonPageData',instanceModel: 'PersonPageType'},
     'PersonList': {dataModel: 'PersonListData',instanceModel: 'PersonListType'},
     'ProfileLandingPage': {dataModel: 'ProfileLandingPageData',instanceModel: 'ProfileLandingPageType'},
     'SearchResultPage': {dataModel: 'SearchResultPageData',instanceModel: 'SearchResultPageType'},
     'SearchSettings': {dataModel: 'SearchSettingsData',instanceModel: 'SearchSettingsType'},
+    'ScriptInjectionSettings': {dataModel: 'ScriptInjectionSettingsData',instanceModel: 'ScriptInjectionSettingsType'},
     'ReferencePageSettings': {dataModel: 'ReferencePageSettingsData',instanceModel: 'ReferencePageSettingsType'},
     'LayoutSettings': {dataModel: 'LayoutSettingsData',instanceModel: 'LayoutSettingsType'},
     'LabelSettings': {dataModel: 'LabelSettingsData',instanceModel: 'LabelSettingsType'},
     'CollectionSettings': {dataModel: 'CollectionSettingsData',instanceModel: 'CollectionSettingsType'},
-    'PersonPage': {dataModel: 'PersonPageData',instanceModel: 'PersonPageType'},
     'StandardPage': {dataModel: 'StandardPageData',instanceModel: 'StandardPageType'},
+    'SpaMedia': {dataModel: 'SpaMediaData',instanceModel: 'SpaMediaType'},
     'SpaFolder': {dataModel: 'SpaFolderData',instanceModel: 'SpaFolderType'},
     'CategoryRoot': {dataModel: 'CategoryRootData',instanceModel: 'CategoryRootType'},
     'PdfFile': {dataModel: 'PdfFileData',instanceModel: 'PdfFileType'},
-    'SettingsFolder': {dataModel: 'SettingsFolderData',instanceModel: 'SettingsFolderType'},
     'BreadcrumbBlock': {dataModel: 'BreadcrumbBlockData',instanceModel: 'BreadcrumbBlockType'},
     'CallToActionBlock': {dataModel: 'CallToActionBlockData',instanceModel: 'CallToActionBlockType'},
     'ButtonBlock': {dataModel: 'ButtonBlockData',instanceModel: 'ButtonBlockType'},
+    'ResetPasswordPage': {dataModel: 'ResetPasswordPageData',instanceModel: 'ResetPasswordPageType'},
     'CarouselBlock': {dataModel: 'CarouselBlockData',instanceModel: 'CarouselBlockType'},
-    'SpaMedia': {dataModel: 'SpaMediaData',instanceModel: 'SpaMediaType'},
-    'ContainerBlock': {dataModel: 'ContainerBlockData',instanceModel: 'ContainerBlockType'},
-    'ResetPasswordMailPage': {dataModel: 'ResetPasswordMailPageData',instanceModel: 'ResetPasswordMailPageType'},
     'ProfilePage': {dataModel: 'ProfilePageData',instanceModel: 'ProfilePageType'},
+    'CodingFile': {dataModel: 'CodingFileData',instanceModel: 'CodingFileType'},
+    'SettingsFolder': {dataModel: 'SettingsFolderData',instanceModel: 'SettingsFolderType'},
     'BlogItemPage': {dataModel: 'BlogItemPageData',instanceModel: 'BlogItemPageType'},
     'BlogListPage': {dataModel: 'BlogListPageData',instanceModel: 'BlogListPageType'},
     'StandardCategory': {dataModel: 'StandardCategoryData',instanceModel: 'StandardCategoryType'},
@@ -33,23 +35,23 @@ export default class TypeMapper extends Loaders.BaseTypeMapper {
     'HomePage': {dataModel: 'HomePageData',instanceModel: 'HomePageType'},
     'LandingPage': {dataModel: 'LandingPageData',instanceModel: 'LandingPageType'},
     'ThreeColumnLandingPage': {dataModel: 'ThreeColumnLandingPageData',instanceModel: 'ThreeColumnLandingPageType'},
-    'ResetPasswordPage': {dataModel: 'ResetPasswordPageData',instanceModel: 'ResetPasswordPageType'},
     'TwoColumnLandingPage': {dataModel: 'TwoColumnLandingPageData',instanceModel: 'TwoColumnLandingPageType'},
+    'LocationItemPage': {dataModel: 'LocationItemPageData',instanceModel: 'LocationItemPageType'},
     'LocationListPage': {dataModel: 'LocationListPageData',instanceModel: 'LocationListPageType'},
     'TagPage': {dataModel: 'TagPageData',instanceModel: 'TagPageType'},
+    'WebImageMediaData': {dataModel: 'WebImageMediaDataData',instanceModel: 'WebImageMediaDataType'},
+    'VectorImageMediaData': {dataModel: 'VectorImageMediaDataData',instanceModel: 'VectorImageMediaDataType'},
     'VideoFile': {dataModel: 'VideoFileData',instanceModel: 'VideoFileType'},
     'StandardFile': {dataModel: 'StandardFileData',instanceModel: 'StandardFileType'},
     'ImageMediaData': {dataModel: 'ImageMediaDataData',instanceModel: 'ImageMediaDataType'},
     'FoundationPdfFile': {dataModel: 'FoundationPdfFileData',instanceModel: 'FoundationPdfFileType'},
-    'CodingFile': {dataModel: 'CodingFileData',instanceModel: 'CodingFileType'},
     'BookmarksPage': {dataModel: 'BookmarksPageData',instanceModel: 'BookmarksPageType'},
-    'LocationItemPage': {dataModel: 'LocationItemPageData',instanceModel: 'LocationItemPageType'},
-    'HealthChatbotBlock': {dataModel: 'HealthChatbotBlockData',instanceModel: 'HealthChatbotBlockType'},
     'GoogleMapsBlock': {dataModel: 'GoogleMapsBlockData',instanceModel: 'GoogleMapsBlockType'},
-    'HeroBlockCallout': {dataModel: 'HeroBlockCalloutData',instanceModel: 'HeroBlockCalloutType'},
+    'ContainerBlock': {dataModel: 'ContainerBlockData',instanceModel: 'ContainerBlockType'},
+    'CustomViewConfigurationBlock': {dataModel: 'CustomViewConfigurationBlockData',instanceModel: 'CustomViewConfigurationBlockType'},
+    'HealthChatbotBlock': {dataModel: 'HealthChatbotBlockData',instanceModel: 'HealthChatbotBlockType'},
     'CustomizedSearchSettings': {dataModel: 'CustomizedSearchSettingsData',instanceModel: 'CustomizedSearchSettingsType'},
-    'HeroBlock': {dataModel: 'HeroBlockData',instanceModel: 'HeroBlockType'},
-    'CustomizedSearchAdvancedSettings': {dataModel: 'CustomizedSearchAdvancedSettingsData',instanceModel: 'CustomizedSearchAdvancedSettingsType'},
+    'CustomizedSearchBlock': {dataModel: 'CustomizedSearchBlockData',instanceModel: 'CustomizedSearchBlockType'},
     'FilterActivitiesBlock': {dataModel: 'FilterActivitiesBlockData',instanceModel: 'FilterActivitiesBlockType'},
     'FilterContinentsBlock': {dataModel: 'FilterContinentsBlockData',instanceModel: 'FilterContinentsBlockType'},
     'FilterDistancesBlock': {dataModel: 'FilterDistancesBlockData',instanceModel: 'FilterDistancesBlockType'},
@@ -57,23 +59,24 @@ export default class TypeMapper extends Loaders.BaseTypeMapper {
     'CalendarBlock': {dataModel: 'CalendarBlockData',instanceModel: 'CalendarBlockType'},
     'YouTubeBlock': {dataModel: 'YouTubeBlockData',instanceModel: 'YouTubeBlockType'},
     'VimeoBlock': {dataModel: 'VimeoBlockData',instanceModel: 'VimeoBlockType'},
-    'CustomizedSearchBlock': {dataModel: 'CustomizedSearchBlockData',instanceModel: 'CustomizedSearchBlockType'},
+    'CustomizedSearchAdvancedSettings': {dataModel: 'CustomizedSearchAdvancedSettingsData',instanceModel: 'CustomizedSearchAdvancedSettingsType'},
     'TwitterBlock': {dataModel: 'TwitterBlockData',instanceModel: 'TwitterBlockType'},
-    'VideoBlock': {dataModel: 'VideoBlockData',instanceModel: 'VideoBlockType'},
-    'MenuItemBlock': {dataModel: 'MenuItemBlockData',instanceModel: 'MenuItemBlockType'},
-    'NavigationBlock': {dataModel: 'NavigationBlockData',instanceModel: 'NavigationBlockType'},
-    'FacebookBlock': {dataModel: 'FacebookBlockData',instanceModel: 'FacebookBlockType'},
-    'CustomViewConfigurationBlock': {dataModel: 'CustomViewConfigurationBlockData',instanceModel: 'CustomViewConfigurationBlockType'},
+    'TextBlock': {dataModel: 'TextBlockData',instanceModel: 'TextBlockType'},
+    'TeaserBlock': {dataModel: 'TeaserBlockData',instanceModel: 'TeaserBlockType'},
     'RssReaderBlock': {dataModel: 'RssReaderBlockData',instanceModel: 'RssReaderBlockType'},
     'PageListBlock': {dataModel: 'PageListBlockData',instanceModel: 'PageListBlockType'},
-    'TeaserBlock': {dataModel: 'TeaserBlockData',instanceModel: 'TeaserBlockType'},
-    'TextBlock': {dataModel: 'TextBlockData',instanceModel: 'TextBlockType'},
+    'FacebookBlock': {dataModel: 'FacebookBlockData',instanceModel: 'FacebookBlockType'},
+    'NavigationBlock': {dataModel: 'NavigationBlockData',instanceModel: 'NavigationBlockType'},
+    'MenuItemBlock': {dataModel: 'MenuItemBlockData',instanceModel: 'MenuItemBlockType'},
+    'HeroBlockCallout': {dataModel: 'HeroBlockCalloutData',instanceModel: 'HeroBlockCalloutType'},
+    'VideoBlock': {dataModel: 'VideoBlockData',instanceModel: 'VideoBlockType'},
+    'HeroBlock': {dataModel: 'HeroBlockData',instanceModel: 'HeroBlockType'},
     'TextboxElementBlock': {dataModel: 'TextboxElementBlockData',instanceModel: 'TextboxElementBlockType'},
     'TextareaElementBlock': {dataModel: 'TextareaElementBlockData',instanceModel: 'TextareaElementBlockType'},
     'NumberElementBlock': {dataModel: 'NumberElementBlockData',instanceModel: 'NumberElementBlockType'},
-    'DateTimeElementBlock': {dataModel: 'DateTimeElementBlockData',instanceModel: 'DateTimeElementBlockType'},
-    'DateTimeRangeElementBlock': {dataModel: 'DateTimeRangeElementBlockData',instanceModel: 'DateTimeRangeElementBlockType'},
     'AddressesElementBlock': {dataModel: 'AddressesElementBlockData',instanceModel: 'AddressesElementBlockType'},
+    'DateTimeRangeElementBlock': {dataModel: 'DateTimeRangeElementBlockData',instanceModel: 'DateTimeRangeElementBlockType'},
+    'DateTimeElementBlock': {dataModel: 'DateTimeElementBlockData',instanceModel: 'DateTimeElementBlockType'},
     'RangeElementBlock': {dataModel: 'RangeElementBlockData',instanceModel: 'RangeElementBlockType'},
     'UrlElementBlock': {dataModel: 'UrlElementBlockData',instanceModel: 'UrlElementBlockType'},
     'SelectionElementBlock': {dataModel: 'SelectionElementBlockData',instanceModel: 'SelectionElementBlockType'},
